move layout logic to the respective operation
